Preserve wanted ratio even when no upscaling
This commit adds logic to deal when upscaling is disabled.

The previous implementation simply max values at their original (source)
values. This commit checks for the wanted ratio and adjust both values
so the destination image will never be larger, but the requested ratio
will remain unchanged.

By example, a 1029x683 image used as source, when requesting as
/2/1029/100/5, would get a 1029x100 image. When requesting
/2/1030/100/5, you'll end up with a 1029x99 image, in order to preseve
the 103:1 ratio.

Width is checked first and both values are adjust. The subsequent check
on the height validates the already scale down dst_h value.

Fixes #106