Content-Disposition: attachment for download
This commit refactors existing code to let the developer create an new
endpoint that will force the user to save the file.

This behavior is not automatically alvailable, since it could be not
desired. Also, we could have pass down the whole query string to the
request, but that open all kinds of trouble (like overriding mode).

Finally, this commit reuses some old "dead" code that was present in the
code base. I could not figure out how the extension could trigger the
execution of the command. It was mostly there for extension to use:
extension developers should use the new code base anyhow.