Prevent memory exhaustion with estimate of needs
This commits adds a feature that will try do detect if the server has
enough memory to process the request or not.

After loading the original image, if enabled, JIT will estimate the
needed memory in order to execute the demand, like so:

$dst_w * $dst_h * ($number_of_channels + 1) * $factor

The factor value is taken from the 'memory_exhaustion_factor'
configuration value.

When enabled, any errors will get loggued.

By default, this feature is turned off

Fixes #134
Re #133