Configurable cache invalidation odds
This commits adds a new hidden configuration value, `cache_invalidation_odds`
which allows developer to control how often cache is validated.

This is particularly usefull on stressed servers, and even better if
serving remote images, since it can lead to a 60% request time
diminution.

The default behaviour is not changed and cache validation will always
occur.

It also updates the x-jit debug header to specify the result of the
cache validation

Fixes #138